+++ front/app/containers/SignUpIn/utils/openSignUpInModalIfNecessary.ts
